[FIX] mail: message reaction from emoji picker test keeps reaction
Before this commit, test "Add the same reaction twice from the emoji
picker" may fail non-deterministically in 18.1 on last step: the
message reaction is sometimes kept, sometimes removed.

The feature works in 18.0 but no longer in 18.1. The test was still
passing though very rarely can fail and actually show the feature
doesn't work.

This happens because the test is clicking on reaction too fast and
doesn't assert on UI the 1st message reaction has been added, thus
it may flicker and keep showing there's 1 message reaction still.

This commit fixes the test in 18.0, in preparation to keep the same
test in 18.1 and show the bugged change of behaviour in 18.1 to fix.

Fixes runbot error 229340

closes odoo/odoo#218152

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>