[FIX] pos_loyalty: max usage partner with points
When an existing partner with points tries to claim a reward from a
loyalty program that has reached its max usage in the current session,
he was still able to claim the reward.

Steps to reproduce:
-------------------
* Create a loyalty program with max usage set to 1
* Create a partner and assign points to him
* Open PoS and select any partner and make an order
* Claim the reward from the loyalty program
* Select the partner that already has points
* Make an order and try to claim the reward again
> Observation: The reward can be claimed, even though the max usage
  is already reached.

Why the fix:
------------
When updating programs, we were not removing the couponPointChanges
for programs that are not applicable anymore. We now make sure to delete
them when the program is not applicable anymore.

opw-4805704

closes odoo/odoo#220201

X-original-commit: 00ed2bf27b96de102850e50eda3df8a58b07891e
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>